Changed getValue methods to return a Number instead of a float.
This is a more generic class provided by Java to cover all numbers and it allows IntegerParameter to return an int instead of a float.
Further Parameter types will be able to return any number value.

If it is possible for IParameters to encode Strings or some other type then this will need changing again. Perhaps we could type the Parameters?